Made code work on states
The code is now structure to operate based on the state of the robot. It
currently uses a simple binary state system.
The arc search has been tuned to not do an extra turn.
The robot now looks at the closest points first if it was just on the
line.